requirements.txt
google-cloud-storage==2.10.0  
google-cloud-bigquery==3.9.0 
apache-airflow[google]==2.10.5